feat: migrate browsers
