+++ index.html
@@ -75,8 +75,8 @@
                     <td><a href="/maintenance/patch_db.php?pwd=secret">Patch DB</a></td>
                 </tr>
                 <tr action>
-                    <td><a href="http://www.cryptomedic.org/api/-api.version-/admin/cron">cron</a></td>
-                    <td><a href="/api/-api.version-/admin/cron">cron</a></td>
+                    <td><a href="http://www.cryptomedic.org/api/-api.version-/browsers/clean">browsers clean</a></td>
+                    <td><a href="/api/-api.version-/browsers/clean">browsers clean</a></td>
                 </tr>
                 <tr action>
                     <td></td>
@@ -89,9 +89,9 @@
                     <td><a href="/api/-api.version-/admin/database">database-structure</a></td>
                 </tr>
                 <tr>
-                    <td><a href="http://www.cryptomedic.org/api/-api.version-/routes/struct/browsers.php">browsers</a>
+                    <td><a href="http://www.cryptomedic.org/api/-api.version-/browsers/show">browsers</a>
                     </td>
-                    <td><a href="/api/-api.version-/routes/struct/browsers.php">browsers</a></td>
+                    <td><a href="/api/-api.version-/routes/browsers/show">browsers</a></td>
                 </tr>
                 <tr>
                     <td><a href="http://www.cryptomedic.org/maintenance/md5sum.php?filter=remote">md5sum remote</a></td>
